Project cleanup: UIImage+Flattened extension, tweak InfoAlert, // MARK
