+++ Experiences/Image/ImageProcessor.swift
@@ -24,15 +24,3 @@ struct ImageProcessor {
         return UIImage(cgImage: renderedCGImage)
     }
 }
-
-
-extension UIImage {
-    
-    /// Renders the image if the pixel data was rotated due to orientation of camera
-    var flattened: UIImage {
-        if imageOrientation == .up { return self }
-        return UIGraphicsImageRenderer(size: size, format: imageRendererFormat).image { context in
-            draw(at: .zero)
-        }
-    }
-}
